Changed default behavior. Added ability to remove pagination.
I realized that the default behavior should just be to mirror the
superset collection, until you actually set a `perPage` parameter.
I added tests and an implementation for this.

Also adds a `removePagination` method to help with this.

Also optimizes the onAddRemove function which was a bottleneck for
collections with a large `perPage` (~30000 in Firefox on my machine).